refactor tests into sub-modules
